+++ models/v4_modify_cluster_request_workers_items_memory.go
@@ -16,7 +16,7 @@ import (
 type V4ModifyClusterRequestWorkersItemsMemory struct {
 
 	// RAM size in GB. Can be an integer or float.
-	SizeGb float64 `json:"size_gb,omitempty"`
+	SizeGb float32 `json:"size_gb,omitempty"`
 }
 
 // Validate validates this v4 modify cluster request workers items memory
